fix: Make reads in `read_int` aligned
This may (or may not) fix some performance issues on ARM, which
generally deals extremely poorly with unaligned reads (see #7). As far
as I am able to test, there is negligible performance impact for this on
x86, likely because the average number of reads per call hasn't changed
much.

Another approach which was taken was attempting to call
copy_nonoverlapping to copy byte-by-byte to a u64 buffer. This was
significantly slower than the previous implementation, and thus dropped.

Fixes: #7